signing/requirements.txt
wmi
pywin32
pycryptodome